update
change env variable in social admin env to social_admin removed var dump for account type for home controller added check if variable exists in registration form to prevent error messages to show up as admin user changed in auth type to account type and value to 1 1 is a value for admin user in our database removed int conversion for checking for admin in config php

+++ Web_application/core/Config.php
@@ -17,7 +17,7 @@ class Config
     {
         if ($_SERVER['APP_ENV'] === '' || $_SERVER['APP_ENV'] !== 'social_admin') {
             $_SERVER['APP_ENV'] = 'regular';
-        }else if((int)Auth::isAdmin()===1){
+        }else if(Auth::isAdmin()===1){
              $_SERVER['APP_ENV'] = 'social_admin';
         }
     }
